requirements.txt
neumai
google-cloud-aiplatform
pandas
google-cloud-storage
unstructured[all-docs]==0.12.2
unstructured[local-inference]==0.12.2
poppler-utils
pillow
psycopg2
sqlalchemy
google-cloud-documentai-toolbox
google-cloud-documentai
fitz
celery
redis
gevent
gcsfs==2023.6.0
flask
PyPDF2
flower
llama-index>=0.9.44
google-cloud-monitoring
streamlit>=1.26.0
langchain>=0.0.217
trubrics>=1.4.3
streamlit-feedback
requests
google-generativeai
google-ai-generativelanguage
llama_hub
spacy
datasets
span-marker
scikit-learn
sqlalchemy[asyncio]
pgvector
psycopg2-binary
asyncpg
wandb>=0.16.2
ragas
nest-asyncio
llama_index
llama-index-embeddings-Gemini
llama-index-vector-stores-postgres
llama-index-embeddings-Gemini
llama-index-callbacks-wandb
llama-index-readers-file
llama-index-llms-Vertex
llmsherpa
llama-index-llms-anthropic